Aligning MetricsResource, extracting DTOs, cleaning interface.
