+++ graylog2-rest-models/src/main/java/org/graylog2/rest/models/system/metrics/requests/MetricNamesResponse.java
@@ -1,0 +1,18 @@
+package org.graylog2.rest.models.system.metrics.requests;
+
+import com.fasterxml.jackson.annotation.JsonCreator;
+import com.fasterxml.jackson.annotation.JsonProperty;
+import com.google.auto.value.AutoValue;
+
+import java.util.Set;
+
+@AutoValue
+public abstract class MetricNamesResponse {
+    @JsonProperty
+    public abstract Set<String> names();
+
+    @JsonCreator
+    public static MetricNamesResponse create(@JsonProperty("names") Set<String> names) {
+        return new AutoValue_MetricNamesResponse(names);
+    }
+}
