+++ graylog2-rest-models/src/main/java/org/graylog2/rest/models/system/metrics/requests/MetricsReadRequest.java
@@ -14,7 +14,7 @@
  * You should have received a copy of the GNU General Public License
  * along with Graylog2.  If not, see <http://www.gnu.org/licenses/>.
  */
-package org.graylog2.rest.resources.system.requests;
+package org.graylog2.rest.models.system.metrics.requests;
 
 import com.fasterxml.jackson.annotation.JsonAutoDetect;
 import com.fasterxml.jackson.annotation.JsonCreator;
