+++ graylog2-server/src/main/java/org/graylog2/rest/resources/system/GrokResource.java
@@ -26,8 +26,8 @@ import org.graylog2.grok.GrokPattern;
 import org.graylog2.grok.GrokPatternService;
 import org.graylog2.plugin.database.ValidationException;
 import org.graylog2.rest.resources.system.responses.GrokPatternList;
-import org.graylog2.security.RestPermissions;
 import org.graylog2.shared.rest.resources.RestResource;
+import org.graylog2.shared.security.RestPermissions;
 
 import javax.inject.Inject;
 import javax.validation.Valid;
