+++ graylog2-shared/src/main/java/org/graylog2/shared/metrics/MetricUtils.java
@@ -14,7 +14,7 @@
  * You should have received a copy of the GNU General Public License
  * along with Graylog2.  If not, see <http://www.gnu.org/licenses/>.
  */
-package org.graylog2.metrics;
+package org.graylog2.shared.metrics;
 
 import com.codahale.metrics.Histogram;
 import com.codahale.metrics.Meter;
